Refactored. Tutor focussed.
