+++ globals.py
@@ -1,4 +1,2 @@
 # globals.py
-
 input_future = None
-initiate_chat_task_created = False
